BREAKING: Expose JWT validation, change validateAuthorizationHeader
This commit exposes JWT validation as a utility, but in doing so also introduces a
**breaking change** that allows `JWTSecuredRequest.validateAuthorizationHeader` to return
multiple errors.

The format before was:

```
{ isValid: true }
// or:
{ isValid: false, msg: 'User-visible message', err: 'More details for developer' }
```

The format now is:

```
{ errors: [], token: {} }; // (token is the decoded JWT object)
// or:
{
   errors: [ { msg: 'User-visible message', err: 'More details for developer' } ],
   token: {},
}
```

In this new format: there may be multiple errors; Also, token is the decoded JWT, and is
undefined if the JWT could not be decoded.