BREAKING: Change error format and methods, provide APIError class
In our APIs, we are standardizing on a consistent error format. This commit adds support
for returning JSON API responses using that error format, as well as a new APIError class
that is exposed so it can be used for creating errors.

ALSO BREAKING: the methods used on ResponseBuilder to add errors have changed
significantly and you will need to update your code to use them correctly when you upgrade
to this version.

The schema of an error message is in example-schemas/error-message.json. It is expected
that an error response returns an array of these (schema at
example-schemas/error-messages-list.json) as its body.

ALSO: Introduces a new peer dependency on the `uuid` library.