Fix "console.error is not a function" on newer versions of node
See https://travis-ci.org/silvermine/apigateway-utils/builds/463798309?utm_medium=notification&utm_source=email

The problem was that when we used sinon to override `console.log` in the
tests, the global one was impacted, causing the error when disparate
parts of the code used other `console` functions. So in this commit,
rather than overriding `console` with sinon, we override a local `log`
function that would normally be bound to `console.log`.