Refs #22 Mostly revert "Fix "console.error is not a function" on newer versions of node"
This reverts the log changes from commit 89d9eaeadc9b3208553b9724acf89474f4fec662. As
noted in #22, the binding of the `console.log` function on code initialization causes
Lambda to log the incorrect `requestID` most of the time. The proper way to fix this would
be to add a mockable logging library. However, that is out of scope at this time.
Therefore, this commit removes the mock of `console.log` in the unit tests. Not the best
for the developer, but removes the risk of unit test debug messages getting swallowed by
an improperly mocked global `console.log`.